Update README.md
Changed a few things. 

Added links, mention Elm in Elm once more and add its contributor to the contributor list, as well as fixed some sentences and introduced Elm like spacing in order to improve the readability of the Readme.md source. 

Line "In case of any questions about the project feel free to submit them in Issues with Q&A label" could create a new issue for the user - with q&a label already applied. Has to be done by priviliged person.